refactor(shared): move logos to shared
