+++ src/app/shared/logos/logo.component.scss
